Extra stabilisation work for more correct ewmh.client-list use.
The window arrangement logic is far clearer.

This change will minimise struts as well as the non-active windows when a fullscreen app is in the foreground and restore them to normal in the reverse case.

_NET_WORKAREA is updated too.

Unfortunately this doesn't solve SDL apps and firefox only going "fullscreen" within the strut-constrained _NET_WORKAREA.

Thankfully mpv goes completely fullscreen and that's probably the most important for me at the moment.