Simplify window resizing and display code.
It turns out most of the ConfigureRequest and ConfigureNotify resizing isn't needed.

The stuff that duplicated this on the MapRequest side is good enough.

ie, results are equally as shoddy as before but with less code!

These display updates are still a work in progress as i've identified other potential areas of improvement.