Redo the window activation and banish code.
This fixes the bug introduced that failed to handle deleting the last window in a desktop.

A new categorise-client-windows function is added and most of the drawing routines have been updated to use it.

Window selection is now much more robust for cases like moving in and out of desktops as well as resizing already mapped windows.